move ssl file to examples/ssl.
